+++ src/Responsive.php
@@ -5,15 +5,20 @@ namespace JustBetter\GlideDirective;
 use Illuminate\Contracts\View\Factory;
 use Illuminate\Contracts\View\View;
 use Statamic\Assets\Asset;
+use Statamic\Fields\Value;
 use Statamic\Statamic;
 
 class Responsive
 {
-    public static function handle(mixed ...$arguments): Factory|View
+    public static function handle(mixed ...$arguments): Factory|View|string
     {
         $image = $arguments[0];
-        $image = get_class($image) === 'Statamic\Fields\Value' ? $image->value() : $image;
+        $image = $image instanceof Value ? $image->value() : $image;
         $arguments = $arguments[1] ?? [];
+
+        if (! $image || ! ($image instanceof Asset)) {
+            return '';
+        }
 
         return view('statamic-glide-directive::image', [
             'image' => $image,
@@ -45,7 +50,8 @@ class Responsive
         }
 
         $configPresets = self::getPresetsByRatio($image, $config);
-
+        $imageMeta = $image->meta();
+        $fit = isset($imageMeta['data']['focus']) ? sprintf('crop-%s', $imageMeta['data']['focus']) : null;
         $index = 0;
         foreach ($configPresets as $preset => $data) {
             $size = $data['w'].'w';
@@ -55,21 +61,21 @@ class Responsive
             }
 
             if (self::canUseWebpSource()) {
-                $glideUrl = Statamic::tag($preset === 'placeholder' ? 'glide:data_url' : 'glide')->params(['preset' => $preset, 'src' => $image->url(), 'format' => 'webp', 'fit' => $data['crop'] ?? 'crop_focal'])->fetch();
+                $glideUrl = Statamic::tag($preset === 'placeholder' ? 'glide:data_url' : 'glide')->params(['preset' => $preset, 'src' => $image->url(), 'format' => 'webp', 'fit' => $fit ?? $data['fit']])->fetch();
                 if ($glideUrl) {
                     $presets['webp'] .= $glideUrl.' '.$size;
                 }
             }
 
             if (self::canUseMimeTypeSource()) {
-                $glideUrl = Statamic::tag($preset === 'placeholder' ? 'glide:data_url' : 'glide')->params(['preset' => $preset, 'src' => $image->url(), 'fit' => $data['crop'] ?? 'crop_focal'])->fetch();
+                $glideUrl = Statamic::tag($preset === 'placeholder' ? 'glide:data_url' : 'glide')->params(['preset' => $preset, 'src' => $image->url(), 'fit' => $fit ?? $data['fit']])->fetch();
                 if ($glideUrl) {
                     $presets[$image->mimeType()] .= $glideUrl.' '.$size;
                 }
             }
 
             if ($preset === 'placeholder') {
-                $glideUrl = Statamic::tag('glide:data_url')->params(['preset' => 'placeholder', 'src' => $image->url(), 'fit' => $data['crop'] ?? 'crop_focal'])->fetch();
+                $glideUrl = Statamic::tag('glide:data_url')->params(['preset' => 'placeholder', 'src' => $image->url(), 'fit' => $fit ?? $data['fit']])->fetch();
                 if ($glideUrl) {
                     $presets['placeholder'] = $glideUrl;
                 }
@@ -88,19 +94,11 @@ class Responsive
 
     protected static function getPresetsByRatio(Asset $image, array $config): array
     {
-        $ratio = $image->width() / $image->height();
         $presets = collect($config);
 
         // filter config based on aspect ratio
-        // if ratio < 1 get all presets with a height bigger than the width, else get all presets with width equal or bigger than the height.
-        if ($ratio < 0.999) {
-            $presets = $presets->filter(fn ($preset, $key) => $preset['h'] > $preset['w']);
-            if ($presets->isNotEmpty() && isset($config['placeholder'])) {
-                $presets->prepend($config['placeholder'], 'placeholder');
-            }
-        } else {
-            $presets = $presets->filter(fn ($preset, $key) => $key === 'placeholder' || $preset['w'] >= $preset['h']);
-        }
+        $vertical = $image->height() > $image->width();
+        $presets = $presets->filter(fn ($preset, $key) => $key === 'placeholder' || (($preset['h'] > $preset['w']) === $vertical));
 
         return $presets->isNotEmpty() ? $presets->toArray() : $config;
     }